@@ -1,13 +1,15 @@
-import React from 'react'
-import '../styles/App.css';
-const App = () => {
+import React from "react";
+import Home from "./pages/Home";
+import {Routes, Route} from "react-router-dom";
 
-
+function App() {
   return (
-    <div id="main">
-    </div>
-  )
+      <>
+      <Routes>
+        <Route path="*" element={<Home />} />
+      </Routes>
+      </>
+  );
 }
-
 
 export default App;
